Enable multiple CUDA computes to be detected and enabled
* Now enables multiple computes
* Removes CUDA_DETECTED_COMPUTE cmake variable. Using COMPUTES_DETECTED_LIST
* Still disable detection using CUDA_DETECT_COMPUTE=OFF
* Setting COMPUTES_DETECTED_LIST="20;30;xy" etc will disable computes detection
  and enable the specified computes
* Can also manually set CUDA_COMPUTE_XY to ON
