+++ src/backend/cuda/CMakeLists.txt
@@ -8,10 +8,22 @@ INCLUDE("${CMAKE_MODULE_PATH}/FindNVVM.cmake")
 
 # Disables running cuda_compute_check.c when build windows using remote
 OPTION(CUDA_COMPUTE_DETECT "Run autodetection of CUDA Architecture" ON)
-IF(CUDA_COMPUTE_DETECT AND NOT DEFINED CUDA_DETECTED_COMPUTE)
+IF(CUDA_COMPUTE_DETECT AND NOT DEFINED COMPUTES_DETECTED_LIST)
     INCLUDE("${CMAKE_MODULE_PATH}/CUDACheckCompute.cmake")
-    SET(CUDA_COMPUTE_${CUDA_DETECTED_COMPUTE} ON CACHE BOOL "" FORCE)
 ENDIF()
+
+LIST(LENGTH COMPUTES_DETECTED_LIST COMPUTES_LEN)
+IF(${COMPUTES_LEN} EQUAL 0)
+    MESSAGE(STATUS "No computes detected. Fall back to 20, 30, 50")
+    LIST(APPEND COMPUTES_DETECTED_LIST "20" "30" "50")
+ENDIF()
+LIST(LENGTH COMPUTES_DETECTED_LIST COMPUTES_LEN)
+MESSAGE(STATUS "Number of Computes = ${COMPUTES_LEN}")
+
+FOREACH(COMPUTE_DETECTED ${COMPUTES_DETECTED_LIST})
+    MESSAGE(STATUS "Setting Compute ${COMPUTE_DETECTED} to ON")
+    SET(CUDA_COMPUTE_${COMPUTE_DETECTED} ON CACHE BOOL "" FORCE)
+ENDFOREACH()
 
 OPTION(CUDA_COMPUTE_20 "CUDA Compute Capability 2.0" OFF)
 OPTION(CUDA_COMPUTE_30 "CUDA Compute Capability 3.0" OFF)
@@ -158,7 +170,7 @@ SOURCE_GROUP(api\\cpp\\Sources FILES ${cpp_sources})
 
 LIST(LENGTH COMPUTE_VERSIONS COMPUTE_COUNT)
 IF(${COMPUTE_COUNT} EQUAL 1)
-  SET(CUDA_NVCC_FLAGS "${CUDA_NVCC_FLAGS} ${CUDA_GENERATE_CODE}")
+    SET(CUDA_NVCC_FLAGS "${CUDA_NVCC_FLAGS} ${CUDA_GENERATE_CODE}")
 ELSE()
     SET(CUDA_NVCC_FLAGS "${CUDA_NVCC_FLAGS} -arch sm_20")
 ENDIF()
